feat: removed tasknotifier somewhat elegantly maintaining DRY
I struggled this for more than 2 hours, but the core thing I was trying
to accomplish was to remove the tasknotifier from the streaming actors
with the least number of changes.

This turned out to be a lot harder than I thought. Root of the problem
was that the `as const` that I was doign in envelopeType.ts to
automatically build a registry broke when I tried to use the same const
lists to derive another list that would be the type for StreamingActors
alone. Typescript does not like this and makes it hard.

So i went line by line and figired that useAvailableActors hook in the
message store could be used to ignore certain actors in the tabs. This
solved my problem somewhat.

However I am now aware of a few new problems:
- the name streamingActors is misleading, it could be actors that stream
  a response but also actors that deserve a tab. I am using it
  interchangeably.
- The tasknotifier even though not a streaming tab takes over the
  priority of the manager tab.
- The manager tab is streaming two blocks, why I need to investigate.