refactor: improve separation of concerns between endpoint and authmanager
The AuthManager class was returning responses in the format that the
endpoint liked which is bad programming. You want those concerns to be
separated.

Seaprated them out by returning exceptions when corner cases are
encountered, and the endpoint decides how it wants to handle the
exceptions and its own contract with the client.